+++ CHANGELOG.md
@@ -295,3 +295,4 @@
 - 2024-02-16: Builder README, leads, contacts, task, +3
 - 2024-02-17: Integ deal, mod, users, find email, +6
 - 2024-02-18: Wire README, frontier, discovery, mutate, +3
+- 2024-02-18: Pool frontier, users, bettercontact, follow up, +5